* Flytter skalSkriveFritekstGrunnetFastsettingAvBeregning inn i VedtakProsessStegInitPanel
